pyproject.toml
[project]
name = "for-my-gf"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aiogram>=3.20.0.post0",
    "alembic>=1.15.2",
    "asyncpg>=0.30.0",
    "fastapi>=0.115.12",
    "itsdangerous>=2.2.0",
    "loguru>=0.7.3",
    "miniopy-async>=1.22.1",
    "pre-commit>=4.2.0",
    "pydantic-settings>=2.9.1",
    "python-multipart>=0.0.20",
    "redis>=6.1.0",
    "sqladmin>=0.20.1",
    "sqlalchemy>=2.0.40",
    "taskiq>=0.11.17",
    "uvicorn>=0.34.2",
    "wemake-python-styleguide>=1.1.0",
]
